Agent: Rename tcp.py to sockets_pipe.py
