+++ monkey/infection_monkey/network/relay/__init__.py
@@ -1,5 +1,5 @@
 from .relay_connection_handler import RelayConnectionHandler
 from .relay_user_handler import RelayUser, RelayUserHandler
+from .sockets_pipe import SocketsPipe
 from .tcp_connection_handler import TCPConnectionHandler
-from .tcp import SocketsPipe
 from .tcp_pipe_spawner import TCPPipeSpawner
