+++ monkey/infection_monkey/network/relay/sockets_pipe.py
